Добавлено сохранение

+++ Model/Field.cs
@@ -6,13 +6,14 @@ using System.Threading.Tasks;
 
 namespace Winding
 {
-    class Field: Build_area
-    {       
-        private string field_coord="";
+
+    class Field : Build_area
+    {
+        private string field_coord = "";
         private double y_up = 0;//верхний край области построения
         private double y_down = 3000;//пока так, но потом это должно тоже считаться 
 
-        public Field(double Area_width, double Field_quantity) : base(Area_width, Field_quantity) { }       
+        public Field(double Area_width, int Field_quantity) : base(Area_width, Field_quantity) { }       
 
         public string Paint_field()
         {
